requirements.txt
bs4==0.0.2
google-search-results==2.4.2
langchain_chroma==0.2.2
langchain_community==0.3.21
langchain_openai==0.3.10
pandas==2.2.3
pypdf==5.4.0
python-dotenv==1.1.0
scholarly==1.7.11
scidownl==1.0.2
Crawl4AI==0.5.0.post8
pydantic==2.11.2